Upgrades to compute_drift
This is quite a reworking, which tries to be more robust about
how it selects a peak to fit, and more verbose when it's uncertain.

Peak finding is now done as follows: find local maximum that are
greater than (mean + .75(max - mean)). If there's only one of these,
and it's within rmax_shift of the 1st order prediction for the center,
take it. If there's more than one, choose the one with the highest
value (and display all the peaks). If there aren't any, do an
initial fit, in a big box around x_pred.